fix: Used function declaration on naming the components, rename jsx files and format codes
